LPS-66935 Move asset publisher data handlers to the proper package
