+++ modules/apps/web-experience/asset/asset-publisher-web/src/main/java/com/liferay/asset/publisher/web/internal/exportimport/data/handler/RelatedAssetsPortletDataHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.asset.publisher.web.internal.lar;
+package com.liferay.asset.publisher.web.internal.exportimport.data.handler;
 
 import com.liferay.asset.publisher.web.constants.AssetPublisherPortletKeys;
 import com.liferay.exportimport.kernel.lar.PortletDataHandler;
